Moved JBakeException into app package
