+++ src/main/java/org/jbake/app/Oven.java
@@ -16,7 +16,6 @@ import com.orientechnologies.orient.core.record.impl.ODocument;
 
 import org.apache.commons.configuration.CompositeConfiguration;
 import org.jbake.app.ConfigUtil.Keys;
-import org.jbake.launcher.JBakeException;
 import org.jbake.model.DocumentTypes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
